refactor: decompose processor.js
