+++ src/main/js/steps/clean.js
@@ -1,0 +1,7 @@
+import {unsetUserConfig} from '../api/git.js'
+import {npmRestore} from '../api/npm.js'
+
+export const clean = async (cwd, packages) => {
+  await unsetUserConfig(cwd)
+  await Promise.all(Object.values(packages).map(npmRestore))
+}
